Checkout: share refund-window selection helper across sidebar and trust cards
The trust card had a getRefundWindowDays() function explicitly described
in a comment as a mirror of the day-count selection inside
CheckoutSummaryRefundWindows. Both surfaces have to agree on which window
to pick (longest plan window for renewal-only / monthly-bundle carts,
shortest otherwise) and which carts get no refund line at all
(non-refundable, all-domain). Refund rules changing in one place and not
the other was a real drift risk.

Add getRefundWindowSummary( cart ) to refund-policies.tsx returning
{ days, usePlanProductName } | null. The flag captures the predicate the
sidebar uses to render "X-day money back guarantee for %(product)s" so
the predicate logic itself isn't recomputed by the consumer; the trust
card ignores the flag because its copy is product-agnostic.

Drops the local mirror function from CheckoutTrustCards and the inline
day-count + predicate branching from CheckoutSummaryRefundWindows.
