+++ client/my-sites/checkout/src/components/checkout-summary-refund-windows.tsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import { Icon, reusableBlock } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
 import { CheckIcon } from './check-icon';
-import { getRefundPolicies, getRefundWindows, RefundPolicy } from './refund-policies';
+import { getRefundWindowSummary } from './refund-policies';
 import type { TranslateResult } from 'i18n-calypso';
 
 const StyledIcon = styled( Icon )`
@@ -48,89 +48,30 @@ export function CheckoutSummaryRefundWindows( {
 } ) {
 	const translate = useTranslate();
 
-	const refundPolicies = getRefundPolicies( cart );
-	const refundWindows = getRefundWindows( refundPolicies );
-
-	if ( ! refundWindows.length || refundPolicies.includes( RefundPolicy.NonRefundable ) ) {
+	const summary = getRefundWindowSummary( cart );
+	if ( ! summary ) {
 		return null;
 	}
-
-	const allCartItemsAreDomains = refundPolicies.every(
-		( refundPolicy ) =>
-			refundPolicy === RefundPolicy.DomainNameRegistration ||
-			refundPolicy === RefundPolicy.DomainNameRegistrationBundled ||
-			refundPolicy === RefundPolicy.DomainNameRenewal
-	);
-
-	if ( allCartItemsAreDomains ) {
-		return null;
-	}
-
-	const allCartItemsAreMonthlyPlanBundle = refundPolicies.every(
-		( refundPolicy ) =>
-			refundPolicy === RefundPolicy.DomainNameRegistration ||
-			refundPolicy === RefundPolicy.PlanMonthlyBundle
-	);
-
-	const allCartItemsArePlanOrDomainRenewals = refundPolicies.every(
-		( refundPolicy ) =>
-			refundPolicy === RefundPolicy.DomainNameRenewal ||
-			refundPolicy === RefundPolicy.PlanMonthlyRenewal ||
-			refundPolicy === RefundPolicy.PlanYearlyRenewal ||
-			refundPolicy === RefundPolicy.PlanBiennialRenewal
-	);
+	const { days, usePlanProductName } = summary;
 
 	let text: TranslateResult;
-
-	if ( refundWindows.length === 1 ) {
-		const refundWindow = refundWindows[ 0 ];
-		const planBundleRefundPolicy = refundPolicies.find(
-			( refundPolicy ) =>
-				refundPolicy === RefundPolicy.PlanBiennialBundle ||
-				refundPolicy === RefundPolicy.PlanYearlyBundle
-		);
+	if ( usePlanProductName ) {
 		const planProduct = cart.products.find( isPlan );
-
-		if ( planBundleRefundPolicy ) {
-			text = translate(
-				'%(days)d-day money back guarantee for %(product)s',
-				'%(days)d-day money back guarantee for %(product)s',
-				{
-					count: refundWindow,
-					args: {
-						days: refundWindow,
-						product: planProduct?.product_name ?? '',
-					},
-				}
-			);
-		} else {
-			text = translate( '%(days)d-day money back guarantee', '%(days)d-day money back guarantee', {
-				count: refundWindow,
-				args: { days: refundWindow },
-			} );
-		}
-	} else if ( allCartItemsAreMonthlyPlanBundle || allCartItemsArePlanOrDomainRenewals ) {
-		const refundWindow = Math.max( ...refundWindows );
-		const planProduct = cart.products.find( isPlan );
-
 		text = translate(
 			'%(days)d-day money back guarantee for %(product)s',
 			'%(days)d-day money back guarantee for %(product)s',
 			{
-				count: refundWindow,
+				count: days,
 				args: {
-					days: refundWindow,
+					days,
 					product: planProduct?.product_name ?? '',
 				},
 			}
 		);
 	} else {
-		const shortestRefundWindow = Math.min( ...refundWindows );
-
 		text = translate( '%(days)d-day money back guarantee', '%(days)d-day money back guarantee', {
-			count: shortestRefundWindow,
-			args: { days: shortestRefundWindow },
-			comment: 'The number of days until the shortest refund window in the cart expires.',
+			count: days,
+			args: { days },
 		} );
 	}
