+++ client/my-sites/checkout/src/components/checkout-trust-cards.tsx
@@ -10,7 +10,7 @@ import styled from '@emotion/styled';
 import { Icon } from '@wordpress/components';
 import { reusableBlock, shield, payment } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
-import { getRefundPolicies, getRefundWindows, RefundPolicy } from './refund-policies';
+import { getRefundWindowSummary } from './refund-policies';
 
 const TrustCardsRow = styled.div`
 	display: grid;
@@ -72,55 +72,9 @@ const CardLogos = styled.div`
 	}
 `;
 
-// Mirrors the day-count selection in `CheckoutSummaryRefundWindows`
-// (wp-checkout-order-summary.tsx) so the trust card and sidebar always
-// agree on the headline number — for example, a yearly plan with a
-// separately-purchased domain shows 4 days in both, while an all-renewals
-// cart shows the longer plan window in both.
-function getRefundWindowDays( cart: ResponseCart ): number | null {
-	const refundPolicies = getRefundPolicies( cart );
-	const refundWindows = getRefundWindows( refundPolicies );
-
-	if ( ! refundWindows.length || refundPolicies.includes( RefundPolicy.NonRefundable ) ) {
-		return null;
-	}
-
-	const allCartItemsAreDomains = refundPolicies.every(
-		( refundPolicy ) =>
-			refundPolicy === RefundPolicy.DomainNameRegistration ||
-			refundPolicy === RefundPolicy.DomainNameRegistrationBundled ||
-			refundPolicy === RefundPolicy.DomainNameRenewal
-	);
-	if ( allCartItemsAreDomains ) {
-		return null;
-	}
-
-	if ( refundWindows.length === 1 ) {
-		return refundWindows[ 0 ];
-	}
-
-	const allCartItemsAreMonthlyPlanBundle = refundPolicies.every(
-		( refundPolicy ) =>
-			refundPolicy === RefundPolicy.DomainNameRegistration ||
-			refundPolicy === RefundPolicy.PlanMonthlyBundle
-	);
-	const allCartItemsArePlanOrDomainRenewals = refundPolicies.every(
-		( refundPolicy ) =>
-			refundPolicy === RefundPolicy.DomainNameRenewal ||
-			refundPolicy === RefundPolicy.PlanMonthlyRenewal ||
-			refundPolicy === RefundPolicy.PlanYearlyRenewal ||
-			refundPolicy === RefundPolicy.PlanBiennialRenewal
-	);
-	if ( allCartItemsAreMonthlyPlanBundle || allCartItemsArePlanOrDomainRenewals ) {
-		return Math.max( ...refundWindows );
-	}
-
-	return Math.min( ...refundWindows );
-}
-
 export default function CheckoutTrustCards( { cart }: { cart: ResponseCart } ) {
 	const translate = useTranslate();
-	const refundDays = getRefundWindowDays( cart );
+	const refundDays = getRefundWindowSummary( cart )?.days ?? null;
 
 	return (
 		<TrustCardsRow className="checkout-trust-cards">
